+++ sdk/clientcore/http-netty4/src/test/java/io/clientcore/http/netty4/implementation/Netty4ResponseHandlerTests.java
@@ -16,8 +16,9 @@ import io.netty.handler.codec.http.LastHttpContent;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
 
+import java.io.IOException;
+import java.io.UncheckedIOException;
 import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ThreadLocalRandom;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReference;
 
@@ -66,98 +67,87 @@ public class Netty4ResponseHandlerTests {
     }
 
     @Test
-    public void incompleteIgnoredResponseBody() {
-        byte[] ignoredBodyData = new byte[32];
-        ThreadLocalRandom.current().nextBytes(ignoredBodyData);
+    public void incompleteIgnoredResponseBody() throws InterruptedException {
+        CountDownLatch headersLatch = new CountDownLatch(1);
+        Netty4ResponseHandler responseHandler = new Netty4ResponseHandler(new HttpRequest().setMethod(HttpMethod.HEAD),
+            new AtomicReference<>(), new AtomicReference<>(), headersLatch);
 
-        HttpRequest request = new HttpRequest().setMethod(HttpMethod.HEAD);
-        AtomicReference<ResponseStateInfo> responseReference = new AtomicReference<>();
-        CountDownLatch latch = new CountDownLatch(1);
-
-        Netty4ResponseHandler responseHandler
-            = new Netty4ResponseHandler(request, responseReference, new AtomicReference<>(), latch);
+        CountDownLatch bodyLatch = new CountDownLatch(1);
 
         Channel ch = createChannelWithReadHandling((readCount, channel) -> {
-            if (readCount == 0) {
-                Netty4ResponseHandler handler = channel.pipeline().get(Netty4ResponseHandler.class);
-                MockChannelHandlerContext ctx = new MockChannelHandlerContext(channel);
-                try {
-                    handler.channelRead(ctx,
-                        new DefaultHttpResponse(HttpVersion.HTTP_1_1, HttpResponseStatus.OK, new DefaultHttpHeaders()));
-                    handler.channelReadComplete(ctx);
-                } catch (Exception ex) {
-                    ctx.fireExceptionCaught(ex);
+            try {
+                if (readCount == 0) {
+                    responseHandler.channelRead(new MockChannelHandlerContext(channel),
+                        new DefaultHttpResponse(HttpVersion.HTTP_1_1, HttpResponseStatus.OK));
+                    responseHandler.channelReadComplete(new MockChannelHandlerContext(channel));
+                } else {
+                    Netty4EagerConsumeChannelHandler eagerConsumer
+                        = channel.pipeline().get(Netty4EagerConsumeChannelHandler.class);
+                    eagerConsumer.channelRead(new MockChannelHandlerContext(channel),
+                        LastHttpContent.EMPTY_LAST_CONTENT);
+                    eagerConsumer.channelReadComplete(new MockChannelHandlerContext(channel));
                 }
-            } else {
-                Netty4EagerConsumeChannelHandler handler
-                    = channel.pipeline().get(Netty4EagerConsumeChannelHandler.class);
-                MockChannelHandlerContext ctx = new MockChannelHandlerContext(channel);
-                handler.channelRead(ctx, Unpooled.wrappedBuffer(ignoredBodyData));
-                handler.channelRead(ctx, Unpooled.wrappedBuffer(ignoredBodyData));
-                handler.channelRead(ctx, Unpooled.wrappedBuffer(ignoredBodyData));
-                handler.channelRead(ctx, Unpooled.wrappedBuffer(ignoredBodyData));
-                handler.channelRead(ctx, LastHttpContent.EMPTY_LAST_CONTENT);
-                handler.channelReadComplete(ctx);
+            } catch (Exception e) {
+                channel.pipeline().fireExceptionCaught(e);
             }
         });
 
         ch.pipeline().addLast(responseHandler);
+
         ch.read();
+        assertTrue(headersLatch.await(10, TimeUnit.SECONDS));
 
-        assertEquals(0, latch.getCount());
+        ch.pipeline().addLast(new Netty4EagerConsumeChannelHandler(bodyLatch, ignored -> {
+        }, false));
 
-        ResponseStateInfo info = responseReference.get();
-        assertNotNull(info);
+        ch.read();
+        assertTrue(bodyLatch.await(10, TimeUnit.SECONDS));
     }
 
     @Test
-    public void bufferedResponseBodyLargerThanInitialRead() {
-        byte[] bodyPieces = new byte[32];
-        ThreadLocalRandom.current().nextBytes(bodyPieces);
-
-        byte[] expectedBody = new byte[bodyPieces.length * 4];
-        System.arraycopy(bodyPieces, 0, expectedBody, 0, bodyPieces.length);
-        System.arraycopy(bodyPieces, 0, expectedBody, bodyPieces.length, bodyPieces.length);
-        System.arraycopy(bodyPieces, 0, expectedBody, bodyPieces.length * 2, bodyPieces.length);
-        System.arraycopy(bodyPieces, 0, expectedBody, bodyPieces.length * 3, bodyPieces.length);
-
-        HttpRequest request = new HttpRequest();
+    public void bufferedResponseBodyLargerThanInitialRead() throws InterruptedException {
         AtomicReference<ResponseStateInfo> responseReference = new AtomicReference<>();
-        CountDownLatch latch = new CountDownLatch(1);
+        CountDownLatch headersLatch = new CountDownLatch(1);
 
         Netty4ResponseHandler responseHandler
-            = new Netty4ResponseHandler(request, responseReference, new AtomicReference<>(), latch);
+            = new Netty4ResponseHandler(new HttpRequest(), responseReference, new AtomicReference<>(), headersLatch);
+
+        CountDownLatch bodyLatch = new CountDownLatch(1);
 
         Channel ch = createChannelWithReadHandling((readCount, channel) -> {
-            if (readCount == 0) {
-                Netty4ResponseHandler handler = channel.pipeline().get(Netty4ResponseHandler.class);
-                MockChannelHandlerContext ctx = new MockChannelHandlerContext(channel);
-                try {
-                    handler.channelRead(ctx,
-                        new DefaultHttpResponse(HttpVersion.HTTP_1_1, HttpResponseStatus.OK, new DefaultHttpHeaders()));
-                    handler.channelReadComplete(ctx);
-                } catch (Exception ex) {
-                    ctx.fireExceptionCaught(ex);
+            try {
+                if (readCount == 0) {
+                    responseHandler.channelRead(new MockChannelHandlerContext(channel),
+                        new DefaultHttpResponse(HttpVersion.HTTP_1_1, HttpResponseStatus.OK));
+                    responseHandler.channelReadComplete(new MockChannelHandlerContext(channel));
+                } else {
+                    Netty4EagerConsumeChannelHandler eagerConsumer
+                        = channel.pipeline().get(Netty4EagerConsumeChannelHandler.class);
+                    eagerConsumer.channelRead(new MockChannelHandlerContext(channel),
+                        LastHttpContent.EMPTY_LAST_CONTENT);
+                    eagerConsumer.channelReadComplete(new MockChannelHandlerContext(channel));
                 }
-            } else {
-                Netty4EagerConsumeChannelHandler handler
-                    = channel.pipeline().get(Netty4EagerConsumeChannelHandler.class);
-                MockChannelHandlerContext ctx = new MockChannelHandlerContext(channel);
-                handler.channelRead(ctx, Unpooled.wrappedBuffer(bodyPieces));
-                handler.channelRead(ctx, Unpooled.wrappedBuffer(bodyPieces));
-                handler.channelRead(ctx, Unpooled.wrappedBuffer(bodyPieces));
-                handler.channelRead(ctx, Unpooled.wrappedBuffer(bodyPieces));
-                handler.channelRead(ctx, LastHttpContent.EMPTY_LAST_CONTENT);
-                handler.channelReadComplete(ctx);
+            } catch (Exception e) {
+                channel.pipeline().fireExceptionCaught(e);
             }
         });
 
         ch.pipeline().addLast(responseHandler);
+
         ch.read();
-
-        assertEquals(0, latch.getCount());
-
+        assertTrue(headersLatch.await(10, TimeUnit.SECONDS));
         ResponseStateInfo info = responseReference.get();
         assertNotNull(info);
+
+        ch.pipeline().addLast(new Netty4EagerConsumeChannelHandler(bodyLatch, buf -> {
+            try {
+                buf.readBytes(info.getEagerContent(), buf.readableBytes());
+            } catch (IOException ex) {
+                throw new UncheckedIOException(ex);
+            }
+        }, false));
+
+        ch.read();
+        assertTrue(bodyLatch.await(10, TimeUnit.SECONDS));
     }
 }