+++ sdk/cosmos/azure-cosmos-tests/src/test/java/com/azure/cosmos/rx/ContainerCreateDeleteWithSameNameTest.java
@@ -116,7 +116,7 @@ public class ContainerCreateDeleteWithSameNameTest extends TestSuiteBase {
         };
     }
 
-    @Test(groups = {"emulator"}, dataProvider = "containerRecreateFeedArgProvider", timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, dataProvider = "containerRecreateFeedArgProvider", timeOut = TIMEOUT)
     public <T> void query(
         int ruBeforeDelete,
         String pkPathBeforeDelete,
@@ -160,7 +160,7 @@ public class ContainerCreateDeleteWithSameNameTest extends TestSuiteBase {
                                         .filter(diagnostics -> containsCollectionRefresh(diagnostics, container))
                                         .collect(Collectors.toList());
 
-                                assertThat(feedResponseDiagnostics.size()).isEqualTo(1);
+                                assertThat(feedResponseDiagnostics.size()).isGreaterThanOrEqualTo(1);
                             }
                         }
                     );
@@ -180,7 +180,7 @@ public class ContainerCreateDeleteWithSameNameTest extends TestSuiteBase {
             getPkAfterRecreate);
     }
 
-    @Test(groups = {"emulator"}, dataProvider = "containerRecreateArgProvider", timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, dataProvider = "containerRecreateArgProvider", timeOut = TIMEOUT)
     public <T> void readItem(
         int ruBeforeDelete,
         String pkPathBeforeDelete,
@@ -230,7 +230,7 @@ public class ContainerCreateDeleteWithSameNameTest extends TestSuiteBase {
             getPkAfterRecreate);
     }
 
-    @Test(groups = {"emulator"}, dataProvider = "containerRecreateArgProvider", timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, dataProvider = "containerRecreateArgProvider", timeOut = TIMEOUT)
     public <T> void deleteItem(
         int ruBeforeDelete,
         String pkPathBeforeDelete,
@@ -279,7 +279,7 @@ public class ContainerCreateDeleteWithSameNameTest extends TestSuiteBase {
             pkPathAfterRecreate,
             getPkAfterRecreate);    }
 
-    @Test(groups = {"emulator"}, dataProvider = "containerRecreateArgProvider", timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, dataProvider = "containerRecreateArgProvider", timeOut = TIMEOUT)
     public <T> void upsertItem(
         int ruBeforeDelete,
         String pkPathBeforeDelete,
@@ -329,7 +329,7 @@ public class ContainerCreateDeleteWithSameNameTest extends TestSuiteBase {
             getPkAfterRecreate);
     }
 
-    @Test(groups = {"emulator"}, dataProvider = "containerRecreateArgProvider", timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, dataProvider = "containerRecreateArgProvider", timeOut = TIMEOUT)
     public <T> void createItem(
         int ruBeforeDelete,
         String pkPathBeforeDelete,
@@ -373,7 +373,7 @@ public class ContainerCreateDeleteWithSameNameTest extends TestSuiteBase {
             getPkAfterRecreate);
     }
 
-    @Test(groups = {"emulator"}, dataProvider = "containerRecreateFeedArgProvider", timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, dataProvider = "containerRecreateFeedArgProvider", timeOut = TIMEOUT)
     public <T> void changeFeedProcessor(
         int ruBeforeDelete,
         String pkPathBeforeDelete,
@@ -458,7 +458,7 @@ public class ContainerCreateDeleteWithSameNameTest extends TestSuiteBase {
             pkPathAfterRecreate);
     }
 
-    @Test(groups = {"emulator"}, dataProvider = "containerRecreateArgProvider", timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, dataProvider = "containerRecreateArgProvider", timeOut = TIMEOUT)
     public <T> void replaceItem(
         int ruBeforeDelete,
         String pkPathBeforeDelete,
@@ -513,7 +513,7 @@ public class ContainerCreateDeleteWithSameNameTest extends TestSuiteBase {
             getPkAfterRecreate);
     }
 
-    @Test(groups = {"emulator"}, dataProvider = "containerRecreateArgProvider", timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, dataProvider = "containerRecreateArgProvider", timeOut = TIMEOUT)
     public <T> void patchItem(
         int ruBeforeDelete,
         String pkPathBeforeDelete,
@@ -571,7 +571,7 @@ public class ContainerCreateDeleteWithSameNameTest extends TestSuiteBase {
             getPkAfterRecreate);
     }
 
-    @Test(groups = {"emulator"}, dataProvider = "containerRecreateArgProvider", timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, dataProvider = "containerRecreateArgProvider", timeOut = TIMEOUT)
     public <T> void batch(
         int ruBeforeDelete,
         String pkPathBeforeDelete,
@@ -609,7 +609,7 @@ public class ContainerCreateDeleteWithSameNameTest extends TestSuiteBase {
             getPkAfterRecreate);
     }
 
-    @Test(groups = {"emulator"}, dataProvider = "containerRecreateArgProvider", timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, dataProvider = "containerRecreateArgProvider", timeOut = TIMEOUT)
     public <T> void bulk(
         int ruBeforeDelete,
         String pkPathBeforeDelete,
@@ -651,7 +651,7 @@ public class ContainerCreateDeleteWithSameNameTest extends TestSuiteBase {
             getPkAfterRecreate);
     }
 
-    @Test(groups = {"emulator"}, dataProvider = "containerRecreateArgProvider", timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, dataProvider = "containerRecreateArgProvider", timeOut = TIMEOUT)
     public void getFeedRanges(
         int ruBeforeDelete,
         String pkPathBeforeDelete,
@@ -685,7 +685,7 @@ public class ContainerCreateDeleteWithSameNameTest extends TestSuiteBase {
             getPkAfterRecreate);
     }
 
-    @Test(groups = {"emulator"}, dataProvider = "containerRecreateFeedArgProvider", timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, dataProvider = "containerRecreateFeedArgProvider", timeOut = TIMEOUT)
     public void queryChangeFeed(
         int ruBeforeDelete,
         String pkPathBeforeDelete,
@@ -757,7 +757,7 @@ public class ContainerCreateDeleteWithSameNameTest extends TestSuiteBase {
             getPkAfterRecreate);
     }
 
-    @Test(groups = {"emulator"}, dataProvider = "containerRecreateFeedArgProvider", timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, dataProvider = "containerRecreateFeedArgProvider", timeOut = TIMEOUT)
     public void readMany(
         int ruBeforeDelete,
         String pkPathBeforeDelete,
@@ -822,13 +822,13 @@ public class ContainerCreateDeleteWithSameNameTest extends TestSuiteBase {
             getPkAfterRecreate);
     }
 
-    @BeforeClass(groups = {"emulator"}, timeOut = SETUP_TIMEOUT)
+    @BeforeClass(groups = {"long-emulator"}, timeOut = SETUP_TIMEOUT)
     public void before_ContainerCreateDeleteWithSameNameTest() {
         client = getClientBuilder().buildAsyncClient();
         createdDatabase = createDatabase(client, testDatabaseId);
     }
 
-    @AfterClass(groups = {"emulator"}, timeOut = SETUP_TIMEOUT)
+    @AfterClass(groups = {"long-emulator"}, timeOut = SETUP_TIMEOUT)
     public void after_ContainerCreateDeleteWithSameNameTest() {
         safeDeleteDatabase(createdDatabase);
         safeClose(client);
@@ -955,9 +955,9 @@ public class ContainerCreateDeleteWithSameNameTest extends TestSuiteBase {
         assertThat(container).isNotNull();
 
         String expectedContainerRid = container.read().block().getProperties().getResourceId();
-        if (!BridgeInternal.getMetaDataDiagnosticContext(cosmosDiagnostics).isEmpty()) {
-            return !BridgeInternal
-                .getMetaDataDiagnosticContext(cosmosDiagnostics)
+        MetadataDiagnosticsContext metaDataDiagnosticContext = BridgeInternal.getMetaDataDiagnosticContext(cosmosDiagnostics);
+        if (metaDataDiagnosticContext != null && !metaDataDiagnosticContext.isEmpty()) {
+            return !metaDataDiagnosticContext
                 .metadataDiagnosticList
                 .stream()
                 .filter(metadataDiagnostics -> {