+++ sdk/dataprotection/azure-resourcemanager-dataprotection/src/main/java/com/azure/resourcemanager/dataprotection/models/AzureOperationalStoreParameters.java
@@ -5,6 +5,7 @@
 package com.azure.resourcemanager.dataprotection.models;
 
 import com.azure.core.annotation.Fluent;
+import com.azure.core.util.logging.ClientLogger;
 import com.azure.json.JsonReader;
 import com.azure.json.JsonToken;
 import com.azure.json.JsonWriter;
@@ -77,8 +78,14 @@ public final class AzureOperationalStoreParameters extends DataStoreParameters {
      */
     @Override
     public void validate() {
-        super.validate();
+        if (dataStoreType() == null) {
+            throw LOGGER.atError()
+                .log(new IllegalArgumentException(
+                    "Missing required property dataStoreType in model AzureOperationalStoreParameters"));
+        }
     }
+
+    private static final ClientLogger LOGGER = new ClientLogger(AzureOperationalStoreParameters.class);
 
     /**
      * {@inheritDoc}