+++ sdk/dataprotection/azure-resourcemanager-dataprotection/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.4.0-beta.1 (Unreleased)
+## 1.5.0-beta.1 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,38 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.4.0 (2025-08-15)
+
+- Azure Resource Manager DataProtection client library for Java. This package contains Microsoft Azure SDK for DataProtection Management SDK. Open API 2.0 Specs for Azure Data Protection service. Package tag package-2025-07-01. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
+
+### Features Added
+
+* `models.AksVolumeTypes` was added
+
+* `models.AdlsBlobBackupDatasourceParameters` was added
+
+* `models.ValidateForModifyBackupRequest` was added
+
+#### `models.BackupInstanceResource` was modified
+
+* `validateForModifyBackup(models.ValidateForModifyBackupRequest)` was added
+* `validateForModifyBackup(models.ValidateForModifyBackupRequest,com.azure.core.util.Context)` was added
+
+#### `models.KubernetesClusterBackupDatasourceParameters` was modified
+
+* `includedVolumeTypes()` was added
+* `withIncludedVolumeTypes(java.util.List)` was added
+
+#### `models.ItemPathBasedRestoreCriteria` was modified
+
+* `withRenameTo(java.lang.String)` was added
+* `renameTo()` was added
+
+#### `models.BackupInstances` was modified
+
+* `validateForModifyBackup(java.lang.String,java.lang.String,java.lang.String,models.ValidateForModifyBackupRequest,com.azure.core.util.Context)` was added
+* `validateForModifyBackup(java.lang.String,java.lang.String,java.lang.String,models.ValidateForModifyBackupRequest)` was added
 
 ## 1.3.0 (2024-07-22)